feat(edit): add command (#525)
Using our building blocks, add `downstackEdit` to allow for reordering a branch.

Note: I know we're current missing tests, and support for lots of edge cases such as "what if a user writes a weird edit file, etc"

Testing this command is a little harder than normal due to the prompted file, so I'll work on that downstack. For now we can trust that it's unreleased, has no dependencies on it, and has been lightly manually tested (but we expect to iterate on it).