feat(create_branch): empty commits (#543)
**Context:**
We've heard from users a desire to be able to create a branch without committing, and not get an error.

This change enables that workflow by having branch create support always making a commit, sometimes empty if no changes are staged. This does not mean that graphite supports two branches pointing to the same commit, like what happens if you merge a branch into another. However, it does enable the common case.

I considered adding work to try and delete empty commits when possible, but based on research I think git already does that for us when operations like rebases happen. All tests are still passing as expected, so I figure we can try this for now.