fix: use pr number as source of truth for sync instead of branch name (#598)
This brings us back to the old state of the world where we only sync branches that have been submitted on this machine.  This prevents the issue where branches with the same name as closed PRs would be deleted inadvertently and toward the world where graphite only operates on branches it knows about.

The future direction here is that this will correctly sync branches both submitted and retrieved with `gt get`