feat: allow automatic branch names up to GH's max supported length (#616)
Also cleaned up this function a little, think it reads easier now.  Note that we now always add the date between the prefix and the message -- we used to not add it if the message was nearing max branch length, but now that it's way longer, it seems reasonable to always truncate the message in favor of the prefix and date.