fix: move to one-at-a-time submit flow to fix ds edit bug (#740)
The major change in this PR is to move from
- push all branches
- update all PRs
- push all metadata 

to:
- for PR:
     - push branch
     - update PR
     - push metadata

This is to avoid cases where after flipping dependency trees during a downstack edit, GitHub automagically closes branches mid-submit because a former parent's former child appears to have been merged.

Since the validation returns the list of PRs in a topsorted order (parents then children) and we then submit in this order, this is guaranteed to never result in branches getting closed inadvertently.

I also took the opportunity to clean up the output in the happy path.  I don't think its necessary to print all the Git output.

![Screen Shot 2022-05-09 at 23.37.42.png](https://graphite-user-uploaded-assets.s3.amazonaws.com/t6FDnC5t98Dwamvf2BHI/93e11d18-8e10-4090-bad1-ffd0089069dd/Screen%20Shot%202022-05-09%20at%2023.37.42.png)

(i submitted branches with the same name multiple times for this manual test, that's why they are shown as updated even though we thought we were creating them)