feat: allow auth and user commands to be run outside of a repo (#1019)
Adds a contextLite that we use instead of context when we don't need repo info.  should also make those commands faster.  learning that typescript is pretty good at letting you keep refactors small across the rest of the codebase as long as you have one place that you're ok with hiding the poop