chore(docker): add reproducibility through docker
**Why** is the change needed?

I was struggling with mocha hanging indefinitely on my Ubuntu machine,
something about all workers being busy. My workaround was to run the
tests in a docker container. I'm sharing this since I thought it might
be useful for others.

**How** is the need addressed?

- Add a Dockerfile that sets up the requirements according to the README
- Make docker ignore node-modules
- Add a docker-compose file to run said Dockerfile

Concerns / side-effects of the changes:

This should probably not be merged, but be put in a wiki or something
similar. I thought this was the easiest way to share it, however.