fix: issue with get for untracked branches (#1067)
I'm going to actually start filling in PR descriptions, seems like a good habit. 

**Context:**

A user reported that running `downstack get` failed with a confusing error.  It turned out that the `getParentPrecondition` was failing because we had an incorrect implicit assumption that if the branch we `get` exists locally, it is tracked.

**Changes In This Pull Request:**

This change removes that assumption, allowing untracked branches to be overwritten by `get`, and adds slightly different copy for when Graphite sees a branch that is untracked vs. has a different parent locally.  Pulled out a little helper function to keep the logic to overwrite/abort the same.


**Test Plan:**

I plan on adding cohesive testing for `ds get` upstack