fix: feedback state recreation (#1069)
**Context:**

When testing the debug context flow, I realized that we were accidentally persisting a graphite cache for the repo the command was run from to the temp folder.

**Changes In This Pull Request:**

Just add the correct chdir to fix this, I think we can probably make this flow not require running in a repo but the command currently does multiple things.

**Test Plan:**

tested locally