fix: lingering pr body edit file (#1238)
Because the filename is constant, not creating a new directory each time meant we would fail if someone quit the submit process in the middle of writing their description but then still saved the file.

To fix we create a new dir every time.

Also put cleanup in a finally — that's just good practice.